tsb: uart: change the declare type
change the input param from uint8_t to uint32_t in ua_set_divisor.
the uint8_t value will discard the divisor latch (MS) part,
therefore, need to change the type to avoid this issue.

Testing Done: test on BDB2B, the uart test function
Bug Number: trivial
Review URL: https://reviews.leaflabs.com/r/1461/
Reviewed by: fparent, mbolivar

+++ nuttx/arch/arm/src/tsb/tsb_uart.c
@@ -309,7 +309,7 @@ static void dump_regs(uint32_t base)
  * @param divisor The divisor value.
  * @return None.
  */
-static void ua_set_divisor(uint32_t base, uint8_t divisor)
+static void ua_set_divisor(uint32_t base, uint32_t divisor)
 {
     /* Set DLAB */
     ua_reg_bit_set(base, UA_LCR, UA_DLAB);
